feat: Add new feature for image upload
Add a new feature that enables users to upload images to the application. This feature allows users to select and upload images from their devices.

BREAKING CHANGE: The format of uploaded images has been changed to optimize storage and enhance compatibility. Please ensure that images are uploaded in the new supported format.